requirements.txt
Django==1.10.6
numpy==1.12.0
py==1.4.32
pydicom==0.9.9
Pillow==4.0.0
pytest==3.0.6
virtualenv==15.1.0
